Change window information value type into i32 from f64

+++ src/window.rs
@@ -2,14 +2,14 @@ use crate::Error;
 
 #[derive(Default, Debug)]
 pub struct Size {
-    pub width: f64,
-    pub height: f64,
+    pub width: i32,
+    pub height: i32,
 }
 
 #[derive(Default, Debug)]
 pub struct Position {
-    pub x: f64,
-    pub y: f64,
+    pub x: i32,
+    pub y: i32,
 }
 
 pub struct Window(pub(crate) crate::platform_impl::Window);
